requirements.txt
requests
python-docx
openpyxl
